requirements.txt
Flask~=1.1.1
Flask-Cors~=3.0.8
gunicorn~=20.0.4
netCDF4~=1.5.3
numpy~=1.18.2
python-dotenv~=0.12.0
rq~=1.3.0